% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binpi.R
\name{binpi}
\alias{binpi}
\title{Missing Data imputation with BINPI}
\usage{
binpi(df, ntree = 100)
}
\arguments{
\item{df}{\code{\link{data.frame}} with missing values}

\item{ntree}{number of trees, input for \code{\link{boosting}} function}
}
\value{
a complete \code{\link{data.frame}}
}
\description{
The binpi function implements the Boosted Incremental Non-Parametric Imputation (BINPI) method.
This method is designed for single imputation of missing data, utilizing AdaBoost algorithms and decision trees.
}
\details{
Based on Master's Thesis in Computational Statistics:

\emph{Santone, C. (2024). Iterative Incremental Non-Parametric Imputation Random Forests-based: una proposta di r-package per l'imputazione dei missing data.}
}
\references{
\emph{D'Ambrosio, A., Aria, M., & Siciliano, R. (2012). Accurate tree-based missing data imputation and data fusion within the statistical learning paradigm. Journal of classification, 29, 227-258.}
}
