% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binpi2.R
\name{binpi2}
\alias{binpi2}
\title{Missing Data imputation with a BINPI's update}
\usage{
binpi2(df, ntree = 100, maxit = 10, eps = 0.001)
}
\arguments{
\item{df}{\code{\link{data.frame}} with missing values}

\item{ntree}{number of trees, input for \code{\link{boosting}} function}

\item{maxit}{number of iterations}

\item{eps}{value for the stop criterion}
}
\value{
a complete \code{\link{data.frame}}
}
\description{
The binpi2 function is an enhanced version of the BINPI method, extending its capability to handle datasets entirely composed of incomplete variables.
It employs an incremental approach to impute missing data using decision trees and boosting techniques.
}
\details{
Based on Master's Thesis in Computational Statistics:

\emph{Santone, C. (2024). Iterative Incremental Non-Parametric Imputation Random Forests-based: una proposta di r-package per l'imputazione dei missing data.}
}
\references{
\emph{D'Ambrosio, A., Aria, M., & Siciliano, R. (2012). Accurate tree-based missing data imputation and data fusion within the statistical learning paradigm. Journal of classification, 29, 227-258.}
}
