% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iinpirf.R
\name{iinpirf}
\alias{iinpirf}
\title{Missing Data imputation with IINPIRF}
\usage{
iinpirf(df, ntree = 100, maxit = 10, eps = 0.001)
}
\arguments{
\item{df}{\code{\link{data.frame}} with missing values}

\item{ntree}{number of trees, input for \code{\link{randomForest}} function}

\item{maxit}{number of iterations}

\item{eps}{value for the stop criterion}
}
\value{
a complete \code{\link{data.frame}}
}
\description{
The iinpirf function implements the Iterative Incremental Non-Parametric Imputation Random Forests-based (IINPIRF) algorithm.
This non-parametric method is designed to impute missing data (MAR) in an incremental and then iterative manner,
by using Random Forests for the predictions.
}
\details{
Based on Master's Thesis in Computational Statistics:

\emph{Santone, C. (2024). Iterative Incremental Non-Parametric Imputation Random Forests-based: una proposta di r-package per l'imputazione dei missing data.}
}
\references{
\emph{D'Ambrosio, A., Aria, M., & Siciliano, R. (2012). Accurate tree-based missing data imputation and data fusion within the statistical learning paradigm. Journal of classification, 29, 227-258.}

\emph{Van Buuren, S. (1999). Flexible Multivariate Imputation by MICE. TNO Prevention and Health.}

\emph{Stekhoven, D. J., & Buhlmann, P. (2012). MissForest-non-parametric missing value imputation for mixed-type data. Bioinformatics, 28(1), 112-118.}
}
